下载m3u8优化语法

+++ Normal/toys/downM3u8/1-下载m3u8文件.py
@@ -17,16 +17,24 @@ if not os.path.exists(out_put_dir):
 # m3u8网址
 base_url = 'https://example.com/path/'
 m3u8_name = 'index.m3u8'
+# 继续下载文件名,第一次是空的，但后面的就类似于：xxx.ts
+keep_on_file_name = ""
+now_keep_on = (keep_on_file_name != "")  # 跳过第一次
 # 下载m3u8
-response = requests.get(base_url + m3u8_name, headers=headers)
-with open(out_put_dir + m3u8_name, 'wb') as f:
-    f.write(response.content)
+if not now_keep_on:
+    print(f"下载{m3u8_name}")
+    response = requests.get(base_url + m3u8_name, headers=headers)
+    with open(out_put_dir + m3u8_name, 'wb') as f:
+        f.write(response.content)
+
 
 # 下载key(如果m3u8有加密的话)
-key_name = 'key.dat'
-response2 = requests.get(base_url + key_name, headers=headers)
-with open(out_put_dir + key_name, 'wb') as f:
-    f.write(response2.content)
+# if not now_keep_on:
+#     key_name = 'key.dat'
+#     print(f"下载{key_name}")
+#     response2 = requests.get(base_url + key_name, headers=headers)
+#     with open(out_put_dir + key_name, 'wb') as f:
+#         f.write(response2.content)
 
 
 # 解析m3u8并下载TS文件
@@ -45,26 +53,19 @@ def read_m3u8_to_ts(file_path: str) -> list[str]:
 
 
 # 2024-09-15 04:25:14 加入即使出现异常也会继续下载的功能
-# 继续下载文件名,第一次是空的，但后面的就类似于：xxx.ts
-keep_on_file_name = ""
-now_keep_on = True
 result_list = read_m3u8_to_ts(out_put_dir + m3u8_name)
-# 跳过第一次
-if keep_on_file_name == "":
-    now_keep_on = False
 # 正式逻辑
 while True:
     # noinspection PyBroadException
     try:
-        for ts_name in result_list:
+        for index, ts_name in enumerate(result_list, start=1):
             if now_keep_on:
                 if ts_name != keep_on_file_name:
                     continue
                 else:
                     now_keep_on = False
-            print(f"下载{ts_name}")
-            # 做备份
-            keep_on_file_name = ts_name
+            print(f"({index}/{len(result_list)})下载{ts_name}")
+            keep_on_file_name = ts_name  # 做备份
             # 下载
             response = requests.get(base_url + ts_name,
                                     headers=headers)
